split the mylib.py into 3 files
